refactor(sidebar): gather heartbeat publication and freshness
