+++ crates/rimz/src/sidebar/focus_anchor.rs
@@ -614,7 +614,7 @@ mod tests {
         socket
             .set_read_timeout(Some(Duration::from_secs(1)))
             .expect("set socket timeout");
-        crate::sidebar::write_heartbeat(
+        crate::sidebar::heartbeat::write_heartbeat(
             &runtime,
             runtime.workspace_id.clone(),
             &instance,
