+++ crates/rimz/src/sidebar/heartbeat.rs
@@ -3,12 +3,19 @@
 use std::fs;
 use std::io;
 use std::path::{Path, PathBuf};
+use std::time::{Duration, SystemTime};
 
 use jiff::Timestamp;
 use serde::{Deserialize, Serialize};
 use tracing::debug;
 
+use crate::disk::atomic;
+use crate::disk::paths::RuntimePaths;
 use crate::ids::{MuxName, PaneId, SidebarInstanceId, WorkspaceId};
+
+/// Maximum age of a sidebar heartbeat before launch, election, and wakeup
+/// fanout treat the instance as dead and skip it.
+pub const SIDEBAR_HEARTBEAT_TTL: Duration = Duration::from_secs(5);
 
 // v5: the snapshot view-model carries explicit named-channel identity on agent
 // rows and panes. v4 carried `root_class`, and the worktree-group kind
@@ -131,6 +138,83 @@ pub fn read_current_heartbeats(dir: &Path) -> io::Result<Vec<(PathBuf, SidebarHe
         heartbeats.push((path, heartbeat));
     }
     Ok(heartbeats)
+}
+
+#[derive(Debug, thiserror::Error)]
+#[error("writing sidebar heartbeat {path}: {source}")]
+pub struct HeartbeatWriteErr {
+    pub path: PathBuf,
+    #[source]
+    pub source: atomic::AtomicErr,
+}
+
+/// Write this sidebar instance's liveness heartbeat in-process.
+///
+/// The heartbeat is a runtime liveness file, not store truth, so the renderer
+/// owns it directly rather than forking `rimz sidebar heartbeat` once per tick.
+/// The JSON shape and the atomic temp-then-rename are identical to the CLI path
+/// they replace, so the store wakeup fanout and the launch freshness gate that
+/// read it are unchanged. The heartbeat carries this process's build id when
+/// the running image is readable. The renderer ensures the runtime dirs at
+/// startup, so this only does the write.
+pub fn write_heartbeat(
+    runtime: &RuntimePaths,
+    workspace_id: WorkspaceId,
+    instance_id: &SidebarInstanceId,
+    mux: MuxName,
+    session_name: &str,
+    wakeup_socket: &Path,
+    pane_id: Option<PaneId>,
+) -> Result<(), HeartbeatWriteErr> {
+    let mut heartbeat = SidebarHeartbeat::new(
+        workspace_id,
+        instance_id.clone(),
+        mux,
+        session_name,
+        wakeup_socket.to_path_buf(),
+        pane_id,
+    );
+    heartbeat.build = crate::build_id::current().map(str::to_owned);
+    heartbeat.version = Some(crate::build_id::VERSION.to_owned());
+    let path = runtime.sidebar_heartbeat_path(instance_id);
+    // Cache-class: a heartbeat is disposable liveness, rewritten every beat
+    // and gc-swept when stale — surviving a power cut buys nothing.
+    atomic::write_temp_then_rename_cache(&path, &heartbeat)
+        .map_err(|source| HeartbeatWriteErr { path, source })
+}
+
+/// Every fresh, current-protocol sidebar heartbeat in the workspace runtime dir.
+/// The shared scan behind the launch gate, the runtime election, and the reload
+/// liveness set: a stale mtime, unreadable JSON, or mismatched protocol is
+/// skipped (so an old-build sidebar drops out and reload replaces it).
+pub(crate) fn fresh_sidebar_heartbeats(rt: &RuntimePaths) -> Vec<SidebarHeartbeat> {
+    let heartbeats = match read_current_heartbeats(&rt.heartbeat_dir) {
+        Ok(heartbeats) => heartbeats,
+        Err(err) => {
+            debug!(path = %rt.heartbeat_dir.display(), error = %err, "sidebar heartbeat dir unreadable");
+            return Vec::new();
+        }
+    };
+
+    heartbeats
+        .into_iter()
+        .filter(|(path, _)| mtime_within_ttl(path))
+        .map(|(_, heartbeat)| heartbeat)
+        .collect()
+}
+
+pub(crate) fn mtime_within_ttl(path: &Path) -> bool {
+    let modified = match fs::metadata(path).and_then(|meta| meta.modified()) {
+        Ok(modified) => modified,
+        Err(err) => {
+            debug!(path = %path.display(), error = %err, "sidebar runtime file metadata unreadable");
+            return false;
+        }
+    };
+    match SystemTime::now().duration_since(modified) {
+        Ok(age) => age <= SIDEBAR_HEARTBEAT_TTL,
+        Err(_) => true,
+    }
 }
 
 #[cfg(test)]
