+++ crates/rimz/src/sidebar/tests/mod.rs
@@ -4,7 +4,7 @@ use tempfile::TempDir;
 
 use super::*;
 use crate::ids::{MuxName, SidebarInstanceId, WorkspaceId};
-use crate::sidebar::heartbeat::SIDEBAR_PROTOCOL_VERSION;
+use crate::sidebar::heartbeat::{SIDEBAR_PROTOCOL_VERSION, write_heartbeat};
 
 mod launch;
