+++ crates/rimz/src/sidebar/timing.rs
@@ -263,12 +263,8 @@ pub const LINK_STATS_STALE: Duration = Duration::from_secs(10);
 /// publishers so they do not leave a permanent "remote room" badge behind.
 pub const LINK_STATS_EXPIRE: Duration = Duration::from_secs(120);
 
-/// Maximum age of a sidebar heartbeat before launch, election, and wakeup
-/// fanout treat the instance as dead and skip it.
-pub const SIDEBAR_HEARTBEAT_TTL: Duration = Duration::from_secs(5);
-
 /// How often a renderer re-stamps its heartbeat. 2s keeps two missed writes of
-/// slack under [`SIDEBAR_HEARTBEAT_TTL`] — the same 2.5× ratio
+/// slack under [`SIDEBAR_HEARTBEAT_TTL`](super::heartbeat::SIDEBAR_HEARTBEAT_TTL) — the same 2.5× ratio
 /// [`PRESENCE_STAMP_FRESH`] keeps over the plugin keepalive — while avoiding an
 /// atomic file write for every store-delta fetch in a busy fleet.
 pub const HEARTBEAT_WRITE_INTERVAL: Duration = Duration::from_secs(2);
