+++ crates/rimz/tests/integration/backend/tmux/support.rs
@@ -17,7 +17,8 @@ pub(super) use rimz::mux::{
 };
 pub(super) use rimz::pane::PaneRef;
 pub(super) use rimz::pane::keys::NamedKey;
-pub(super) use rimz::sidebar::{SidebarLaunchOutcome, launch_sidebar_if_needed, write_heartbeat};
+pub(super) use rimz::sidebar::heartbeat::write_heartbeat;
+pub(super) use rimz::sidebar::{SidebarLaunchOutcome, launch_sidebar_if_needed};
 pub(super) use rimz::workspace::WorkspaceResolver;
 pub(super) use tempfile::TempDir;
